requirements.txt
confluent-kafka
confluent-kafka[avro]
avro-python3